refactor: moved edit button to header on tale page; moved submit to header on edit tale page; moved delete button to edit tale page
